backend/requirements.txt
boto3
google-generativeai
Pillow
django
pytest